requirements.txt
asgiref==3.11.1
Django==6.0.2
django-allauth==65.14.3
django-simple-history==3.11.0
pillow==12.1.1
sqlparse==0.5.5
